# go/model_comment.go
/*
 * AwesomeMusic API
 *
 * 音楽サービス AwesomeMusicのWebAPI
 *
 * API version: 0.0.1
 * Contact: adf1985adf@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Comment struct {

	Id int64 `json:"id,omitempty"`

	Comment string `json:"comment,omitempty"`
}