# go/model_post.go
/*
 * AwesomeMusic API
 *
 * 音楽サービス AwesomeMusicのWebAPI
 *
 * API version: 0.0.1
 * Contact: adf1985adf@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Post struct {

	Id int64 `json:"id,omitempty"`

	Url string `json:"url,omitempty"`

	Message string `json:"message,omitempty"`
}